server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
africastalking==1.2.5
python-dotenv==1.0.0
mysql-connector-python==8.1.0
aiohttp==3.9.3
requests==2.31.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
phonenumbers==8.13.20
geopy==2.4.0
SQLAlchemy==2.0.21
flask-restx==1.1.0
flask-admin==1.6.1
flask-sqlalchemy==3.1.1
